+++ Sources/VComponents/Components/Buttons/Capsule Button/VCapsuleButtonLabel.swift
@@ -1,5 +1,5 @@
 //
-//  VSecondaryButtonLabel.swift
+//  VCapsuleButtonLabel.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 2/27/22.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-// MARK: - V Secondary Button Label
+// MARK: - V Capsule Button Label
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-typealias VSecondaryButtonLabel = GenericContent_TitleIconTitleContent
+typealias VCapsuleButtonLabel = GenericContent_TitleIconTitleContent